requirements.txt
requests>=2.31.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
APScheduler>=3.10.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyYAML>=6.0
tenacity>=8.2.0
ratelimit>=2.2.0
python-dotenv>=1.0.0
click>=8.1.0
python-json-logger>=2.0.0
